backport: dash#4714
