+++ src/bench/coin_selection.cpp
@@ -29,7 +29,7 @@ static void addCoin(const CAmount& nValue, const CWallet& wallet, std::vector<st
 static void CoinSelection(benchmark::Bench& bench)
 {
     auto chain = interfaces::MakeChain();
-    const CWallet wallet(*chain, WalletLocation(), WalletDatabase::CreateDummy());
+    const CWallet wallet(*chain, WalletLocation(), CreateDummyWalletDatabase());
     std::vector<std::unique_ptr<CWalletTx>> wtxs;
     LOCK(wallet.cs_wallet);
 
@@ -60,7 +60,7 @@ static void CoinSelection(benchmark::Bench& bench)
 
 typedef std::set<CInputCoin> CoinSet;
 static auto testChain = interfaces::MakeChain();
-static const CWallet testWallet(*testChain, WalletLocation(), WalletDatabase::CreateDummy());
+static const CWallet testWallet(*testChain, WalletLocation(), CreateDummyWalletDatabase());
 std::vector<std::unique_ptr<CWalletTx>> wtxn;
 
 // Copied from src/wallet/test/coinselector_tests.cpp
