+++ src/qt/rpcconsole.h
@@ -104,7 +104,6 @@ public Q_SLOTS:
     void setFontSize(int newSize);
 
     /** Wallet repair options */
-    void walletSalvage();
     void walletRescan1();
     void walletRescan2();
     void walletZaptxes1();
