+++ src/wallet/salvage.h
@@ -1,0 +1,14 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+ // Copyright (c) 2009-2020 The Bitcoin Core developers
+ // Distributed under the MIT software license, see the accompanying
+ // file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+ #ifndef BITCOIN_WALLET_SALVAGE_H
+ #define BITCOIN_WALLET_SALVAGE_H
+
+ #include <fs.h>
+ #include <streams.h>
+
+ bool RecoverDatabaseFile(const fs::path& file_path);
+
+ #endif // BITCOIN_WALLET_SALVAGE_H
