+++ src/wallet/test/wallet_test_fixture.cpp
@@ -7,7 +7,7 @@
 
 WalletTestingSetup::WalletTestingSetup(const std::string& chainName)
     : TestingSetup(chainName),
-      m_wallet(*m_chain, WalletLocation(), WalletDatabase::CreateMock())
+      m_wallet(*m_chain, WalletLocation(), CreateMockWalletDatabase())
 {
     bool fFirstRun;
     m_wallet.LoadWallet(fFirstRun);
