+++ test/functional/wallet_basic.py
@@ -399,8 +399,7 @@ class WalletTest(BitcoinTestFramework):
             '-reindex',
             '-zapwallettxes=1',
             '-zapwallettxes=2',
-            # disabled until issue is fixed: https://github.com/bitcoin/bitcoin/issues/7463
-            # '-salvagewallet',
+
         ]
         chainlimit = 6
         for m in maintenance:
